long int coordinates

+++ include/CubieCube.hpp
@@ -6,7 +6,7 @@
 /*   By: mamartin <mamartin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/21 15:29:37 by user42            #+#    #+#             */
-/*   Updated: 2022/04/23 18:24:05 by mamartin         ###   ########.fr       */
+/*   Updated: 2022/04/23 22:20:17 by mamartin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,17 +33,17 @@ class CubieCube : public ACube
 		// member functions
 		virtual void	move(char face, int factor = 1);
 		// coordinates getters
-		unsigned int	getCornerOriCoord() const;
-		unsigned int	getEdgeOriCoord() const;
-		unsigned int	getCornerPermCoord() const;
-		unsigned int	getEdgePermCoord() const;
-		unsigned int	getUDSliceCoord() const;
+		long	getCornerOriCoord() const;
+		long	getEdgeOriCoord() const;
+		long	getCornerPermCoord() const;
+		long	getEdgePermCoord() const;
+		long	getUDSliceCoord() const;
 		// coordinates setters
-		void			setCornerOriCoord(int coordinate);
-		void			setEdgeOriCoord(int coordinate);
-		void			setCornerPermCoord(int coordinate);
-		void			setEdgePermCoord(int coordinate);
-		void			setUDSliceCoord(int coordinate);
+		void	setCornerOriCoord(long coordinate);
+		void	setEdgeOriCoord(long coordinate);
+		void	setCornerPermCoord(long coordinate);
+		void	setEdgePermCoord(long coordinate);
+		void	setUDSliceCoord(long coordinate);
 		
     private:
